+++ src/webapp/components/multiple-selector/MultipleSelector.tsx
@@ -4,19 +4,17 @@ import {
     createStyles,
     Divider,
     FormControl,
-    InputAdornment,
     InputLabel,
     makeStyles,
     MenuItem,
     Select,
-    TextField,
     Theme,
     useTheme,
 } from "@material-ui/core";
 import { MultipleDropdownProps } from "@eyeseetea/d2-ui-components";
-import { Search as SearchIcon } from "@material-ui/icons";
-import { useBooleanState } from "$/webapp/utils/use-boolean";
 import { DisableableTooltip } from "$/webapp/components/disableable-tooltip/DisableableTooltip";
+import { useMultipleSelector } from "$/webapp/components/multiple-selector/useMultipleSelector";
+import { SearchField } from "$/webapp/components/multiple-selector/SearchField";
 import i18n from "$/utils/i18n";
 import _ from "$/domain/entities/generic/Collection";
 
@@ -37,74 +35,35 @@ export interface MultipleSelectorProps extends MultipleDropdownProps {
 
 export const MultipleSelector: React.FC<MultipleSelectorProps> = React.memo(props => {
     const {
-        items,
-        values,
-        onChange,
         label,
         className,
         name,
         allOption,
         customMenu,
         disabled = false,
-        type = "item",
-        pluralType = "items",
         isSearchable = false,
     } = props;
 
     const classes = useStyles();
     const theme = useTheme();
 
-    const [menuIsOpen, { enable: openMenu, disable: closeMenu }] = useBooleanState(false);
-    const [tooltipIsOpen, { enable: openTooltip, disable: closeTooltip }] = useBooleanState(false);
-    const [filterText, setFilterText] = React.useState("");
+    const {
+        menuIsOpen,
+        tooltipIsOpen,
+        openMenu,
+        closeMenu,
+        openTooltip,
+        closeTooltip,
+        filterText,
+        filter,
+        virtualValues,
+        notifyChange,
+        isAllSelected,
+        helperText,
+        filteredItems,
+    } = useMultipleSelector(props);
+
     const mergedClasses = [className, classes.formControl].join(" ");
-
-    const notifyChange = React.useCallback(
-        (event: React.ChangeEvent<{ value: unknown }>) => {
-            onChange(
-                (event.target.value as string[]).filter(
-                    s => s !== "multiple-selector-void" && s !== undefined && s !== null
-                )
-            );
-        },
-        [onChange]
-    );
-
-    const filter = React.useCallback((event: React.ChangeEvent<HTMLInputElement>) => {
-        setFilterText(event.target.value);
-    }, []);
-
-    const isAllSelected = React.useMemo(
-        () => (allOption ? values.includes(allOption.value) : false),
-        [allOption, values]
-    );
-
-    const selected = React.useMemo(() => {
-        if (isAllSelected && allOption) return [allOption.value];
-        else return _(values).isEmpty() ? ["multiple-selector-void"] : values;
-    }, [allOption, isAllSelected, values]);
-
-    const helperText = React.useMemo(() => {
-        if (allOption && isAllSelected)
-            return `${items.length} ${items.length > 1 ? pluralType : type} ` + i18n.t("selected");
-
-        const selectedCount = selected.filter(
-            v => v !== "multiple-selector-void" && v !== undefined
-        ).length;
-
-        return selectedCount
-            ? `${selectedCount} ${selectedCount > 1 ? pluralType : type} ` + i18n.t("selected")
-            : undefined;
-    }, [allOption, isAllSelected, items.length, pluralType, selected, type]);
-
-    const virtualValues = React.useMemo(
-        () => (customMenu?.onOpen && selected.length > 500 ? selected.slice(0, 10) : selected),
-        [customMenu?.onOpen, selected]
-    );
-
-    const renderItems = React.useMemo(() => {
-        return customMenu?.onOpen && items.length > 500 ? items.slice(0, 10) : items;
-    }, [customMenu?.onOpen, items]);
 
     const MenuProps = React.useMemo(
         () => getMenuProps(isSearchable, 2, isSearchable),
@@ -129,6 +88,7 @@ export const MultipleSelector: React.FC<MultipleSelectorProps> = React.memo(prop
                 <InputLabel htmlFor={name} className={classes.label}>
                     {label}
                 </InputLabel>
+
                 <Select
                     id={name}
                     label={label}
@@ -142,55 +102,37 @@ export const MultipleSelector: React.FC<MultipleSelectorProps> = React.memo(prop
                     autoFocus={!isSearchable}
                     multiple
                 >
+                    {/* Note: Material UI does not like React.Fragment as children inside the Menu component,
+                     * so the code separates the fragments under the same condition to avoid Material UI complaining. */}
                     {isSearchable && (
-                        <>
-                            <Box padding={theme.spacing(0, 3.5, 0, 2)}>
-                                <TextField
-                                    id="standard-basic"
-                                    label="Search"
-                                    margin="dense"
-                                    variant="outlined"
-                                    size="small"
-                                    onChange={filter}
-                                    InputProps={{
-                                        endAdornment: (
-                                            <InputAdornment position="end">
-                                                <SearchIcon color="disabled" />
-                                            </InputAdornment>
-                                        ),
-                                    }}
-                                    onKeyDown={e => {
-                                        // Prevents autoselecting item while typing (default Select behaviour)
-                                        if (e.key !== "Escape") e.stopPropagation();
-                                    }}
-                                    onClick={e => e.stopPropagation()}
-                                    value={filterText}
-                                    autoFocus
-                                    fullWidth
-                                />
-                            </Box>
-                            <Box margin="0.75rem 0">
-                                <Divider />
-                            </Box>
-                        </>
+                        <Box padding={theme.spacing(0, 3.5, 0, 2)}>
+                            <SearchField text={filterText} filter={filter} />
+                        </Box>
                     )}
+
+                    {isSearchable && (
+                        <Box margin="0.75rem 0">
+                            <Divider />
+                        </Box>
+                    )}
+
                     <MenuItem className={classes.hide} value="multiple-selector-void" disabled>
                         {i18n.t("Nothing selected")}
                     </MenuItem>
-                    {/* Material UI does not like React.Fragment as children inside the Menu component */}
+
                     {allOption && <MenuItem value={allOption.value}>{allOption.text}</MenuItem>}
+
                     {allOption && (
                         <Box margin="0.75rem 0">
                             <Divider />
                         </Box>
                     )}
-                    {renderItems
-                        .filter(item => item.text.toLowerCase().includes(filterText.toLowerCase()))
-                        .map(item => (
-                            <MenuItem key={item.value} value={item.value} disabled={isAllSelected}>
-                                {item.text}
-                            </MenuItem>
-                        ))}
+
+                    {filteredItems.map(item => (
+                        <MenuItem key={item.value} value={item.value} disabled={isAllSelected}>
+                            {item.text}
+                        </MenuItem>
+                    ))}
                 </Select>
             </FormControl>
         </DisableableTooltip>
@@ -229,12 +171,6 @@ const useStyles = makeStyles((theme: Theme) =>
     })
 );
 
-const ITEM_HEIGHT = 36; // line-height: 1.5 + padding: 6 * 2
-const ITEM_PADDING_TOP = 8;
-const DIVIDER = 12 + 1 + 12;
-const SEARCH_FIELD = 52;
-const IMAGINARY_PADDING_BOTTOM = 8;
-
 function getMenuProps(autoFocus: boolean, dividers = 0, isSearchable = false) {
     return {
         PaperProps: {
@@ -254,3 +190,9 @@ function getMenuProps(autoFocus: boolean, dividers = 0, isSearchable = false) {
         autoFocus: autoFocus,
     } as const;
 }
+
+const ITEM_HEIGHT = 36; // line-height: 1.5 + padding: 6 * 2
+const ITEM_PADDING_TOP = 8;
+const DIVIDER = 12 + 1 + 12;
+const SEARCH_FIELD = 52;
+const IMAGINARY_PADDING_BOTTOM = 8;
